Added node_type to indexed search
